small bug fixes in game manager
made number of folds getter in gene sequence a int and fixed error in game manager.  plan on having models spawn and fly by initial throw speed

+++ FoldedCSC/Assets/Scripts/Genes/GeneSequence.cs
@@ -40,9 +40,9 @@ public class GeneSequence
         this.paperWidth = geneSequence[5];
     }
 
-    public float GetNumberOfFolds()
+    public int GetNumberOfFolds()
     {
-        return numberOfFolds;
+        return (int)numberOfFolds;
     }
 
     public float GetInitialVelocity()
